[NER-811] Added unit tests with cmocked uart.
This commit aims to be a reference for the usage of the (CMock-style)
mocked uart's driver. The tests' dir shows an example of a test
that uses the mocked uart.

In order to mock the ESP-IDF uart, a patch file and instructions
are available in the 'patches' dir.

This commit won't be used because the mocked uart does not enable
unit tests in this component as it uses the 'uart_read/write_bytes'
functions multiples times internally.