[NER-811] Added serial service to handle with esp-idf uart driver.
In order to be able to test ansic1218's services that uses the
serial uart driver, a 'serial' wrapper was added to provide
modularization of such services.

Specifically, the 'transport' class receives a pointer to the serial
service and it is possible to create a mocked serial in order to
test the 'transport' service.